moved userHandler to handlers folder
